agentic_rag客服系统/requirements.txt
# 核心依赖
langchain==1.2.6
langchain-core==1.2.7
langchain-community==0.4.1
langchain-ollama==1.0.1
langchain-chroma==1.1.0
langgraph==1.0.6
chromadb==1.4.1
ollama==0.6.1

# 文档处理
tiktoken==0.5.2
pypdf==6.6.1
python-docx==1.2.0
PyMuPDF==1.26.7
openpyxl==3.1.5

# Web服务
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21

# 工具库
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
loguru==0.7.2
rich==14.2.0

# 数据处理
pandas==2.3.3
numpy==1.26.4
scikit-learn==1.8.0

# 嵌入模型
sentence-transformers==5.2.0
